requirements.txt
numpy
opencv-python==4.8.0.74
Pillow==10.0.0
mysql-connector-python==8.1.0
Pillow==10.0.0
parsel==1.8.1
requests==2.31.0
python-dotenv==1.0.0
playwright==1.40.0
html-diff==0.4.1
